Feat : 고객 메뉴상세보기 가게시리얼넘버+menuId(#8)
1. 메뉴리스트에서 해당 메뉴에 들어가면 id를 파라미터로 상세페이지를 띄움.

+++ jumun/src/main/java/com/mihak/jumun/menu/MenuService.java
@@ -28,7 +28,7 @@ public class MenuService {
         return newMenu.getId();
     }
 
-    public Optional<Menu> findById(Integer id) {
+    public Optional<Menu> findById(long id) {
         return menuRepository.findById(id);
     }
